LayoutTests/accessibility/ios-simulator/text-marker-validation.html

<!DOCTYPE HTML PUBLIC "-//IETF//DTD HTML//EN">
<html>
<head>
<script src="../../resources/js-test-pre.js"></script>
<script>
var successfullyParsed = false;
</script>
</head>
<body>

<div id="element1">element 1</div>
<div id="element2">element 2</div>
<div id="element3">element 3</div>
<div id="element4">element 4</div>

<p id="description"></p>
<div id="console"></div>

<script>

    description("This test ensures that a text range can be found for an element within a web page. ");

    if (window.accessibilityController) {

        var obj = accessibilityController.accessibleElementById("element3");

        var textPos = obj.elementTextPosition;
        var textLen = obj.elementTextLength;

        var succeeded = textPos == 20 && textLen == 9;
        shouldBe("succeeded", "true");
    }

    successfullyParsed = true;
</script>

<script src="../../resources/js-test-post.js"></script>

</body>
</html>
